chore(photoslist): updated
