+++ src/layouts/index.ts
@@ -1,1 +1,2 @@
+export { default as EditPhotoLayout } from './EditPhotoLayout';
 export { default as InternalPageLayout } from './InternalPageLayout';
