+++ src/pages/user/photo/[id]/edit.tsx
@@ -1,0 +1,9 @@
+import { EditPhotoLayout } from '~/layouts';
+
+function EditPhoto(): JSX.Element {
+	return (
+		<EditPhotoLayout editPage />
+	);
+}
+
+export default EditPhoto;
